--- portafly/src/components/pages/accounts/listing/index.ts
@@ -1,4 +1,0 @@
-export * from './AccountsListingTable'
-export * from './ActionButtonImpersonate'
-export * from './utils'
-export { default as AccountsListingPage } from './AccountsListingPage'
