--- portafly/src/components/pages/accounts/listing/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './tableDataFactory'
